DevopsProject/views.py
from django.shortcuts import render


def home(request):
    return render(request, 'index.html')


def calculate(request):
    if request.method == 'POST':
        num1 = float(request.POST['num1'])
        num2 = float(request.POST['num2'])
        operation = request.POST['operation']

        if operation == 'add':
            result = num1 + num2
        elif operation == 'subtract':
            result = num1 - num2
        elif operation == 'multiply':
            result = num1 * num2
        elif operation == 'divide':
            result = num1 / num2
        else:
            result = 'Invalid operation'

        return render(request, 'index.html', {'result': result})
    else:
        return render(request, 'index.html')
